+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Make IndexLite more compatible with Meilisearch
 - Add group filtering to app search results
 - Upgrade TipTap to v3
+- Add parallel job processing to worker using the `parallel` extension if available
 
 ## 2.12.1 (2025-10-26)